requirements.txt
Django>=5.1,<5.3
psycopg[binary]>=3.2,<4.0
requests>=2.32,<3.0
google-api-python-client>=2.179,<3.0
google-auth>=2.40,<3.0
google-auth-oauthlib>=1.2,<2.0
